[14.0][IMP] contract_sale_generation: Split test and common
