+++ contract_sale_generation/tests/__init__.py
@@ -1,3 +1,1 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from . import test_contract_sale
